include mm_types.h directly
The mm_types.h were imported from tizen libmm-common package
and the devel package of gst-plugins-renderer has the header
file. The correct way is we also add the libmm-common package
like tizen, but it's not necessary in this time because only
the mm_types.h is required for plugin packages.

Change-Id: If0a7217b3f62dc6fa0fe728aebe350cc8c3d2a7e
Signed-off-by: Chanho Park <chanho61.park@samsung.com>