gst-plugins-scaler: add a function of closing gem handle
This patch addes a function to close the gem handle
getiing from source's global name.

In the previous patch, scaler used wrong function to close the gem handle
so it was removed because of the stream off fail issue in decoder part
and the gem object will be deleted in source part.
so it's not supposed to be needed.
but the gem object is deleted when there is no handle opened.
so if scaler get a new handle of the gem object, it should be closed
before the source close the gem.

Change-Id: Ie9b202e56aa91a43216ae94b4ec33d4c19190785
Signed-off-by: Hyejungkwon <cjscld15@nexell.co.kr>